aider: Prevented sorting the top item up and the bottom item down by disabling the sort buttons when the item is at the top or bottom of the list.
# Aider chat conversation:

USER: now prevent sorting the top item up and the bottom item down

ASSISTANT: To prevent sorting the top item up and the bottom item down, we need to disable the sort buttons when the item is at the top or bottom of the list, respectively. We can achieve this by adding a conditional `disabled` attribute to the sort buttons.

Here are the changes that need to be made to the `SpoolList.js` file:

```javascript
frontend\spooler\src\components\SpoolList.js
<<<<<<< SEARCH
                <button
                  onClick={() => handleSort(spool, "up")}
                  className="btn btn-sm btn-secondary mx-1"
                >
=======
                <button
                  onClick={() => handleSort(spool, "up")}
                  className={`btn btn-sm btn-secondary mx-1 ${spools[0].id === spool.id ? 'disabled' : ''}`}
                  disabled={spools[0].id === spool.id}
                >
>>>>>>> REPLACE
```

```javascript
frontend\spooler\src\components\SpoolList.js
<<<<<<< SEARCH
                <button
                  onClick={() => handleSort(spool, "down")}
                  className="btn btn-sm btn-secondary mx-1"
                >
=======
                <button
                  onClick={() => handleSort(spool, "down")}
                  className={`btn btn-sm btn-secondary mx-1 ${spools[spools.length - 1].id === spool.id ? 'disabled' : ''}`}
                  disabled={spools[spools.length - 1].id === spool.id}
                >
>>>>>>> REPLACE
```

These changes add a `disabled` attribute to the "up" button when the `spool` is the first item in the `spools` array and to the "down" button when the `spool` is the last item. The `className` is also conditionally modified to include the 'disabled' class for styling purposes when the button is disabled.