aider: Don't animate when swapping sort order, only on add/remove.
# Aider chat conversation:

USER: now just don't animate when swapping sort order, only on add remove

ASSISTANT: To achieve this, we need to modify the `SpoolList.js` file to ensure that the `CSSTransition` component only animates when items are added or removed, not when their sort order changes. We can do this by using the `appear` prop on the `CSSTransition` component, which controls whether the transition should occur on the initial mount. We will set it to `false` to prevent the animation on initial mount, which includes sorting.

Here's the change that needs to be made to `SpoolList.js`:

```javascript
frontend\spooler\src\components\SpoolList.js
<<<<<<< SEARCH
            <CSSTransition key={`${spool.id}-${spool.sort_order}`} timeout={500} classNames="spool">
=======
            <CSSTransition key={spool.id} timeout={500} classNames="spool" appear={false}>
>>>>>>> REPLACE
```

This change will ensure that the `CSSTransition` component only animates when items are added or removed from the list, as the `key` is now only based on the `spool.id` and not on the `spool.sort_order`. The `appear` prop set to `false` will prevent the animation from triggering on the initial mount of the component, which includes the sorting action.